carts: adopt float note_vol/note_res — drop the (int)(level*7) quantizers
Now that note_vol/note_res take float, the 8 carts that hand-rolled the quantizer
get smooth continuous dynamics for free. Tier 1 (the level IS the gesture):
martenot (the touche d'intensite, was quantized to ~6), glassharmonica (the
swell), bowed (bow energy), musicalsaw (bow speed). Modular: modrack AMP + RES
CV (smooth VCA/resonance off a CV). brass mute-res sweep. Engine rumbles
(trafficjam/trackgen) for consistency. Each just drops (int)(x*7+0.5f) -> x*7.0f.
All compile-checked (frames 2); audio-only change, thumbnails unchanged (re-embed
only). mouthharp deferred — it has a pre-existing formant var/API collision.

+++ tools/carts/brass.c
@@ -153,7 +153,7 @@ static void push_live(void) {
         } else {                                              // closing the bell, swept live (drag it = wah)
             note_filter(h[i], FILTER_LOW);
             note_cutoff(h[i], (int)mute_cut());
-            note_res(h[i], (int)mute_res());
+            note_res(h[i], mute_res());
         }
     }
 }
